Mobile drawer
drawer menu voor mobile
en verdere refactor van Navbar component
