+++ Frontend/theater-laak/src/components/navbar/Navbar.js
@@ -1,4 +1,4 @@
-import { AppBar, IconButton, Toolbar, Typography } from "@mui/material";
+import { AppBar, Button, IconButton, Toolbar, Typography } from "@mui/material";
 import { Stack } from "@mui/system";
 import { Link } from "react-router-dom";
 import MenuIcon from "@mui/icons-material/Menu";
@@ -6,40 +6,16 @@ import { useState, useRef, useEffect } from "react";
 import PersonIcon from "@mui/icons-material/Person";
 import ShoppingCartIcon from "@mui/icons-material/ShoppingCart";
 import AccountPopper from "./AccountPopper";
-import NavLinks from "./NavLinks";
 
+import DrawerComp from "./DrawerComp";
+import AccountButton from "./AccountButton";
 export default function Navbar(props) {
-    const [openAccount, setOpenAccount] = useState(false);
-    const anchorRef = useRef(null);
-
-    const handleToggleAccount = () => {
-        setOpenAccount((prev) => !prev);
-    };
-
-    const handleCloseAccount = (event) => {
-        if (anchorRef.current && anchorRef.current.contains(event.target)) {
-            return;
-        }
-        setOpenAccount(false);
-    };
-
-    function handleListKeyDown(event) {
-        if (event.key === "Tab") {
-            event.preventDefault();
-            setOpenAccount(false);
-        } else if (event.key === "Escape") {
-            setOpenAccount(false);
-        }
-    }
-
-    const prevOpenAccount = useRef(openAccount);
-    useEffect(() => {
-        if (prevOpenAccount.current === true && openAccount === false) {
-            anchorRef.current.focus();
-        }
-
-        prevOpenAccount.current = openAccount;
-    }, [openAccount]);
+    const [openDrawer, setOpenDrawer] = useState(false);
+    const navLinks = [
+        { name: "Home", path: "/" },
+        { name: "Agenda", path: "/agenda" },
+        { name: "Steun ons", path: "/steun-ons" },
+    ];
 
     return (
         <AppBar position="static">
@@ -50,6 +26,7 @@ export default function Navbar(props) {
                     color="inherit"
                     aria-label="logo"
                     sx={{ display: { sm: "none" } }}
+                    onClick={() => setOpenDrawer(true)}
                 >
                     <MenuIcon />
                 </IconButton>
@@ -65,32 +42,28 @@ export default function Navbar(props) {
                     spacing={1.5}
                     sx={{ display: { xs: "none", sm: "block" } }}
                 >
-                    <NavLinks />
+                    {navLinks.map((link) => (
+                        <Button key={link.path} color="inherit" sx={{ fontWeight: 600 }}>
+                            <Link to={link.path}>{link.name}</Link>
+                        </Button>
+                    ))}
                 </Stack>
+
                 <Stack direction={"row"}>
-                    <IconButton
-                        color="inherit"
-                        sx={{ fontWeight: 600 }}
-                        ref={anchorRef}
-                        aria-controls={openAccount ? "composition-menu" : undefined}
-                        aria-expanded={openAccount ? "true" : undefined}
-                        aria-haspopup="true"
-                        onClick={handleToggleAccount}
-                    >
+                    <AccountButton>
                         <PersonIcon />
-                    </IconButton>
+                    </AccountButton>
                     <IconButton color="inherit" sx={{ fontWeight: 600 }}>
                         <Link to="/winkelwagen" style={{ fontSize: 0 }}>
                             <ShoppingCartIcon />
                         </Link>
                     </IconButton>
-                    <AccountPopper
-                        openAccount={openAccount}
-                        anchorRef={anchorRef}
-                        handleCloseAccount={handleCloseAccount}
-                        handleListKeyDown={handleListKeyDown}
-                    />
                 </Stack>
+                <DrawerComp
+                    navLinks={navLinks}
+                    openDrawer={openDrawer}
+                    setOpenDrawer={setOpenDrawer}
+                ></DrawerComp>
             </Toolbar>
         </AppBar>
     );
